Ease the creation of others sub-commands for cli
Internally the cli configuration is done with the Service Provider API.
Now, all classes who implements `CliCommand` and are declared as
implementation for the service provider API are added to the cli automatically.
All current sub-commands implements `CliCommand`.
