+++ src/main/java/fr/insalyon/citi/golo/cli/command/CompilerCommand.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright 2012-2015 Institut National des Sciences Appliquées de Lyon (INSA-Lyon)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.insalyon.citi.golo.cli.command;
+
+import com.beust.jcommander.Parameter;
+import com.beust.jcommander.Parameters;
+import fr.insalyon.citi.golo.cli.command.spi.CliCommand;
+import fr.insalyon.citi.golo.compiler.GoloCompilationException;
+import fr.insalyon.citi.golo.compiler.GoloCompiler;
+
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.util.LinkedList;
+import java.util.List;
+
+@Parameters(commandNames = {"compile"}, commandDescription = "Compiles Golo source files")
+public class CompilerCommand implements CliCommand {
+
+  @Parameter(names = "--output", description = "The compiled classes output directory")
+  String output = ".";
+
+  @Parameter(description = "Golo source files (*.golo)")
+  List<String> sources = new LinkedList<>();
+
+  @Override
+  public void execute() throws Throwable {
+    GoloCompiler compiler = new GoloCompiler();
+    File outputDir = new File(this.output);
+    for (String source : this.sources) {
+      File file = new File(source);
+      try (FileInputStream in = new FileInputStream(file)) {
+        compiler.compileTo(file.getName(), in, outputDir);
+      } catch (IOException e) {
+        System.out.println("[error] " + source + " does not exist or could not be opened.");
+        return;
+      } catch (GoloCompilationException e) {
+        handleCompilationException(e);
+      }
+    }
+  }
+}
