+++ src/main/java/fr/insalyon/citi/golo/cli/command/VersionCommand.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2012-2015 Institut National des Sciences Appliquées de Lyon (INSA-Lyon)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.insalyon.citi.golo.cli.command;
+
+import com.beust.jcommander.Parameter;
+import com.beust.jcommander.Parameters;
+import fr.insalyon.citi.golo.cli.command.spi.CliCommand;
+
+@Parameters(commandNames = {"version"}, commandDescription = "Queries the Golo version")
+public class VersionCommand implements CliCommand {
+
+  @Parameter(names = "--full", description = "Prints the full information details")
+  boolean full = false;
+
+  @Override
+  public void execute() throws Throwable {
+    if (this.full) {
+      System.out.println("Golo version: " + Metadata.VERSION + " (build " + Metadata.TIMESTAMP + ")");
+      System.out.println("JVM version: " + System.getProperty("java.version"));
+    } else {
+      System.out.println(Metadata.VERSION);
+    }
+  }
+}
