+++ src/main/java/fr/insalyon/citi/golo/cli/command/spi/CliCommand.java
@@ -1,0 +1,62 @@
+/*
+ * Copyright 2012-2015 Institut National des Sciences Appliquées de Lyon (INSA-Lyon)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.insalyon.citi.golo.cli.command.spi;
+
+import fr.insalyon.citi.golo.compiler.GoloCompilationException;
+
+import java.io.File;
+import java.lang.invoke.MethodHandle;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.net.URLClassLoader;
+import java.util.List;
+
+import static java.lang.invoke.MethodHandles.publicLookup;
+import static java.lang.invoke.MethodType.methodType;
+
+public interface CliCommand {
+
+  void execute() throws Throwable;
+
+  default URLClassLoader primaryClassLoader(List<String> classpath) throws MalformedURLException {
+    URL[] urls = new URL[classpath.size()];
+    int index = 0;
+    for (String element : classpath) {
+      urls[index] = new File(element).toURI().toURL();
+      index = index + 1;
+    }
+    return new URLClassLoader(urls);
+  }
+
+  default void callRun(Class<?> klass, String[] arguments) throws Throwable {
+    MethodHandle main = publicLookup().findStatic(klass, "main", methodType(void.class, String[].class));
+    main.invoke(arguments);
+  }
+
+  default void handleCompilationException(GoloCompilationException e) {
+    if (e.getMessage() != null) {
+      System.out.println("[error] " + e.getMessage());
+    }
+    if (e.getCause() != null) {
+      System.out.println("[error] " + e.getCause().getMessage());
+    }
+    for (GoloCompilationException.Problem problem : e.getProblems()) {
+      System.out.println("[error] " + problem.getDescription());
+    }
+    System.exit(1);
+  }
+}
